% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishersMethod.R
\name{FishersMethod}
\alias{FishersMethod}
\title{Run Fisher's Method to combine p values}
\usage{
FishersMethod(GOde, GOdm)
}
\description{
Fisher's Method combines p values from different statistical models.
}
