% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGenes.R
\name{getAllGenes}
\alias{getAllGenes}
\title{Retrieve topTable of differentially expressed genes}
\usage{
getAllGenes(DGElist, designMatrix)
}
\arguments{
\item{DGElist}{describe this}

\item{designMatrix}{A model matrix that is the output of the `getDesignMatrix` function}
}
\value{
This tells you what object it returns
}
\description{
Retrieve topTable of differentially expressed genes
}
\details{
This function fits a linear model to the gene expression data and retrieves a `topTable` output that has the genes sorted by p value
}
