% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEgenes.R
\name{getDEgenes}
\alias{getDEgenes}
\title{Retrieve differentially expressed genes}
\usage{
getDEgenes(DGElist, variable, metadata)
}
\arguments{
\item{DGElist}{A DGElist output from `limma`}

\item{variable}{The variable of interest}

\item{metadata}{A dataframe with different variables on the x axis and samples on the y axis.}
}
\value{
A character vector of differentially expressed genes.
}
\description{
Retrieve a character vector with a list of differentially expressed genes (should have ENSEMBL IDs)
}
\details{
This function works in a similar way to that of the `getAllGenes` function in that its fitting a linear model to the gene expression data. Then it runs the `topTable` function and retrieves the genes from that data set, ordered by p value.
}
