% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignMatrix.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Create Design Matrix}
\usage{
getDesignMatrix(variable, metadata)
}
\arguments{
\item{variable}{The selected variable of interest, should be a character string or vector.}

\item{metadata}{A dataframe with different variables on the x axis and samples on the y axis.}
}
\value{
A model matrix that is compatible with `limma` and `InteGRAPE` analyses with information on the current variable of interest
}
\description{
This function is just a quick way to make a design matrix
}
\details{
`getDesignMatrix` creates a design matrix from the given `metadata` and `variable` of interest.

This function defined a formula from the supplied information and creates the matrix directly from that information.
}
