% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMgenes.R
\name{getDMgenes}
\alias{getDMgenes}
\title{Retrieve differentially methylated genes}
\usage{
getDMgenes(DGElist, variable, metadata, allGenes, SeqInfo)
}
\arguments{
\item{DGElist}{A DGElist output from `limma`}

\item{variable}{The variable of interest}

\item{metadata}{A dataframe with different variables on the x axis and samples on the y axis.}

\item{allGenes}{An output with the `topTable` from the `getAllGenes` function}

\item{SeqInfo}{A SeqInfo object containing genomic information on your species of interest. Can either find one on ENSEMBL or create one from an object using `seqinfo` https://www.rdocumentation.org/packages/GenomeInfoDb/versions/1.8.3/topics/seqinfo}
}
\value{
A character vector of differentially expressed genes.
}
\description{
Retrieve a character vector with a list of differentially methylated genes (Should have ENSEMBL IDs)
}
\details{
This function works in a similar way to that of the `getAllGenes` function in that its fitting a linear model to the differential methylation data.

This variation of the workflow uses a Generalised linear fit model and runs a `topTags` function on the genes.
}
