% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGOenrichment.R
\name{getGOenrichment}
\alias{getGOenrichment}
\title{Run a GO enrichment analysis}
\usage{
getGOenrichment(de, EG.GO)
}
\arguments{
\item{de}{list of differentially expressed or differentially methylated genes}

\item{EG.GO}{A dataframe containing the information relating to ENSEMBL Gene IDs mapped to their respective GO terms}
}
\value{
A dataframe containing information on the Gene ID, the GO term and ontology, and the assigned p value from the hypergeometric distribution model (`phyper`).
}
\description{
This function retrieves functionally enriched GO terms
}
\details{
This function uses a hypergeometric distribution model to perform the functional enrichment analysis. Essentially

this model finds the probability of retrieving two genes from the entire `universe` with the same GO term. It finds

the probabilities of finding two of the same, one with and one without, or both without the GO term. From this it

assigns a p value to tells whats the probability of viewing this functional enrichment if the null hypothesis is true.
}
