% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegration.R
\name{getIntegration}
\alias{getIntegration}
\title{Integration of GO terms}
\usage{
getIntegration(de, dm, EG.GO)
}
\arguments{
\item{de}{A character vector of differentially expressed genes, retrieved from the `getDEgenes` function.}

\item{dm}{A character vector of differentially methylated genes, retrieved from the `getDMgenes` function.}

\item{EG.GO}{A data frame containing information of ENSEMBL gene IDs mapped to GO IDs and terms}
}
\description{
Retrieve a dataframe of GO IDs and terms that are enriched for both gene expression and methylation
}
\details{
This function runs an intersect between the gene expression GO terms vs the differential methylation GO terms. Thus finding those enriched for both and integrating the information.
}
