% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaledCPM.R
\name{getScaledCPM}
\alias{getScaledCPM}
\title{Retrieve scaled counts per million}
\usage{
getScaledCPM(DGElist, designMartix, nGenes = 30)
}
\arguments{
\item{DGElist}{A DGElist containing the count and sampling data}

\item{nGenes}{Specify the number of genes you'd like to get data for in the top table at the end}

\item{designMatrix}{A model matrix that is the output of the `getDesignMatrix` function}
}
\value{
A molten dataframe containing scaled cpm values
}
\description{
This function retrieves counts per million that have been scaled to allow more effective visualisation
}
\details{
This function will take the count data from a DGElist and scale it in order to enhance the visualisation appearance, especially when making heatmaps. The scaling method mimics that of the `pheatmap` function.
}
