% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopGenes.R
\name{getTopGenes}
\alias{getTopGenes}
\title{Retrieve topTable of differentially expressed genes}
\usage{
getTopGenes(DGElist, designMatrix)
}
\arguments{
\item{DGElist}{An object of class `DGElist` containing information on gene counts and samples, typical output from the `limma` package}

\item{designMartix}{A model matrix containing information on the metadata, an output from the `getDesignMatrix` function}
}
\value{
A dataframe containing information from the linear fit, ranking DE genes on significance
}
\description{
Retrieve topTable of differentially expressed genes
}
\details{
This function fits a linear model to the gene expression data and retrieves a `topTable` output that has the genes sorted by p value
}
