% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopMethylation.R
\name{getTopMethylation}
\alias{getTopMethylation}
\title{Retrieve differential methylation information}
\usage{
getTopMethylation(DGElist, designMatrix)
}
\arguments{
\item{DGElist}{An object of class `DGElist`, a typical output in the `limma` package.}

\item{designMatrix}{A model matrix that is the output of the `getDesignMatrix` function}
}
\description{
Extracts the top differential methylation tags from a `DGElist`
}
\details{
Differential methylation across a set of samples are fitted with a generalised linear model in the context of a metadata variable and then ranked in order of significance
}
