% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeReads.R
\name{normalizeReads}
\alias{normalizeReads}
\title{Normalisation of reads}
\usage{
normalizeReads(DGElist)
}
\arguments{
\item{DGElist}{An object of class `DGElist`, typically an output of the `limma::DGElist` function}
}
\description{
Normalise the data to account for any variation
}
\details{
Experimental batch effects occur when data has been collected from different experiments, which may lead to variation in the data as a result of those independent experiments. Normalisation of the data will normalise this variation and effectively eliminate any variance from batch effects.
}
