% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Multi-dimensional scaling of samples}
\usage{
plot_MDS(DGElist, variable, metadata, size = 80)
}
\arguments{
\item{DGElist}{An object of class `DGElist`, conatining information about counts and samples}

\item{variable}{A character string of the variable under investigation, should be found in the metadata}

\item{metadata}{A data frame containing a set of variables and information relating to them}

\item{size}{Alter the size of the points, 80 is the standard default size}
}
\description{
Plot the clustering of samples to find any similarities or dissimilarities
}
\details{
A multi-dimensional scaling plots clusters the data in an unsuperivsed manner, meaning the data input isn't trained with an expected outcome. An example of a method using a trained data set is a Principle Component Analysis. Multi-dimensional scaling is an efficient way to try and observe if there is any differences between samples before executing any formal analyses.
}
