% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOterms.R
\name{plotGOterms}
\alias{plotGOterms}
\title{Plot significant GO terms}
\usage{
plotGOterms(de, dm, EG.GO)
}
\arguments{
\item{de}{A character vector of differentially expressed genes, an output from the `getDEgenes` function}

\item{dm}{A character vector of differentially methylated genes, an output from the `getDMgenes` function}

\item{EG.GO}{A dataframe containing the information relating to ENSEMBL Gene IDs mapped to their respective GO terms}
}
\value{
A plot showing which GO terms are likely to be enriched
}
\description{
This is a good check to see if we can expect any functionally enriched ontologies before running a formal analysis
}
\details{
Plotting the GO terms associated with differentially expressed genes and differentially methylated genes will give an overview on what to expect when running further GO analyses such as a Fishers Exact Test or Hypergeometric Distribution Model.
}
