% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(DGElist, variable, metadata, nGenes = 30)
}
\arguments{
\item{DGElist}{A DGElist containing the count and sampling data}

\item{variable}{The selected variable of interest, should be a character string or vector.}

\item{metadata}{A dataframe with different variables on the x axis and samples on the y axis.}

\item{nGenes}{Specify the number of genes you'd like to get data for in the top table at the end}
}
\value{
A lovely looking heatmap which is interactive
}
\description{
This function plots a heatmap for direct visualisation of results
}
\details{
This function will plot a heatmap directly from the count data, an annotation bar at the top of the heatmap will offer information about the plot at a glance. A side bar indicating the pvalue will allow determination of statistical significance at a glance as well.
}
