% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinearExpression.R
\name{plotLinearExpression}
\alias{plotLinearExpression}
\title{Retrieve differentially methylated genes}
\usage{
plotLinearExpression(DGElist, designMatrix, allGenes, selectGene = 1)
}
\arguments{
\item{DGElist}{A DGElist output from `limma`}

\item{designMatrix}{A model matrix with information on the variable of interest, the output of `getDesignMatrix`}

\item{allGenes}{An output with the `topTable` from the `getAllGenes` function}

\item{selectGene}{This is a subset operator, it takes an integer or a character vector of a gene ID of interest}
}
\value{
A plot visualising the linear model of gene expression.
}
\description{
Retrieve a character vector with a list of differentially methylated genes (Should have ENSEMBL IDs)
}
\details{
This function fits a linear model to the differential expression data and visualises that model in a scatter plot.
}
