% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattamGeneExpr.R
\name{plotManhattanGeneExpr}
\alias{plotManhattanGeneExpr}
\title{Plot gene positions with differential expression}
\usage{
plotManhattanGeneExpr(topGenes, SeqInfo, Chromosome = "chr1")
}
\arguments{
\item{topGenes}{A dataframe output from the `getTopGenes` function}

\item{SeqInfo}{An object of class `Seqinfo` containing information on the chromosomes within an organism}

\item{Chromosome}{A character string containing the chromosome to plot, ie "chr1", "chr4", "chr13", etc.}
}
\description{
This function plots the results of a differential gene expression analysis within a Manhattan Plot.
}
\details{
Plotting the information regarding postions of differentially expressed genes allows a whole genome view of gene expression and also makes possible a comparison by visualisation of sites with differential methylation.
}
