% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattanMethylation.R
\name{plotManhattanMethylation}
\alias{plotManhattanMethylation}
\title{Plot Methylation positions}
\usage{
plotManhattanMethylation(topMethylation, SeqInfo, Chromosome = "chr1")
}
\arguments{
\item{topMethylation}{A dataframe contianing information on the significance of differential methylation sites, an output from the `getTopMethylation` function}

\item{SeqInfo}{An object of class `Seqinfo` containing information on the chromosomes within an organism}

\item{Chromosome}{A character string containing the chromosome to plot, ie "chr1", "chr4", "chr13", etc.}
}
\description{
This function plots the results of a differential methylation analysis within a Manhattan Plot.
}
\details{
Plotting the information regarding methylation positions allows a whole genome view of methylation and also makes possible a comparison by visualisation of sites for differential gene expression.
}
