% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingHeatmap.R
\name{plottingHeatmap}
\alias{plottingHeatmap}
\title{Plot heatmap}
\usage{
plottingHeatmap(ScaledCPM, variable, metadata)
}
\arguments{
\item{ScaledCPM}{A DGElist containing the scaled count data, the output from the `getScaledCPM` function}

\item{variable}{The selected variable of interest, should be a character string or vector.}

\item{metadata}{A dataframe with different variables on the x axis and samples on the y axis.}
}
\value{
A lovely looking heatmap which is interactive
}
\description{
This function plots a heatmap from the scaled cpm values
}
\details{
This function will plot a heatmap directly from the count data, an annotation bar at the top of the heatmap will offer information about the plot at a glance. A side bar indicating the pvalue will allow determination of statistical significance at a glance as well.
}
