% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariableGO.R
\name{plotVariableGO}
\alias{plotVariableGO}
\title{Plot GO terms from count data}
\usage{
plotVariableGO(DGE_RNAseq, DGE_SubsetMethylation, appmetaG, appmetaM, variable,
  VvSeqInfo, EG.GO)
}
\arguments{
\item{DGE_RNAseq}{A DGElist output from `limma` including data from differential gene expression}

\item{DGE_SubsetMethylation}{A DGElist output from `limma` including data from differential gene methylation}

\item{appmetaG}{A dataframe with different variables on the x axis and samples on the y axis. Specifically for the gene expression data}

\item{appmetaM}{A dataframe with different variables on the x axis and samples on the y axis. Specifically for the gene methylation data}

\item{variable}{The variable of interest, a character string or vector}

\item{VvSeqInfo}{A SeqInfo object containing genomic information on your species of interest. Can either find one on ENSEMBL or create one from an object using `seqinfo` https://www.rdocumentation.org/packages/GenomeInfoDb/versions/1.8.3/topics/seqinfo}

\item{EG.GO}{A dataframe containing the information relating to ENSEMBL Gene IDs mapped to their respective GO terms}
}
\description{
This function allows the plotting of GO terms directly from the two DGElists
}
\details{
This function has the same output as the `plotGOterms` function, but it runs the entire analysis directly from the DGElists. The body will contain the functions used, and it may be better to run the analysis from each step as this function requires several inputs.
}
