% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGenes.R
\name{getAllGenes}
\alias{getAllGenes}
\title{Retrieve topTable of differentially expressed genes}
\usage{
getAllGenes(DGElist, variable, metadata)
}
\arguments{
\item{DGElist}{describe this}

\item{variable}{the variable of interest}

\item{metadata}{describe}
}
\value{
This tells you what object it returns
}
\description{
Retrieve topTable of differentially expressed genes
}
\details{
Paragraph that describes what the function is doing
}
