% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaledCPM.R
\name{getScaledCPM}
\alias{getScaledCPM}
\title{Retrieve scaled counts per million}
\usage{
getScaledCPM(DGElist, variable, metadata, nGenes = 30)
}
\arguments{
\item{DGElist}{A DGElist containing the count and sampling data}

\item{variable}{The selected variable of interest, should be a character string or vector.}

\item{metadata}{A dataframe with different variables on the x axis and samples on the y axis.}

\item{nGenes}{Specify the number of genes you'd like to get data for in the top table at the end}
}
\value{
A molten dataframe containing scaled cpm values
}
\description{
This function retrieves counts per million that have been scaled to allow more effective visualisation
}
\details{
This function will take the count data from a DGElist and scale it in order to enhance the visualisation appearance, especially when making heatmaps. The scaling method mimics that of the `pheatmap` function.
}
