% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopGenes.R
\name{getTopGenes}
\alias{getTopGenes}
\title{Retrieve topTable of differentially expressed genes}
\usage{
getTopGenes(DGElist, variable, metadata)
}
\arguments{
\item{DGElist}{describe this}

\item{variable}{the variable of interest}

\item{metadata}{describe}
}
\value{
This tells you what object it returns
}
\description{
Retrieve topTable of differentially expressed genes
}
\details{
This function fits a linear model to the gene expression data and retrieves a `topTable` output that has the genes sorted by p value
}
